+++ activerecord/lib/active_record/database_configurations/database_config.rb
@@ -21,6 +21,10 @@ module ActiveRecord
         raise NotImplementedError
       end
 
+      def adapter
+        config["adapter"]
+      end
+
       def url_config?
         false
       end